Create a single temporary file for each plantuml file
It does not create the source-file-name.png in the working directory.
Instead, it creates a temporary file in the temporary directory so that
the generated png file is not shown in the git's commit list.
TODO: The default SLT image viewer seems not to support zooming. Any
good alternative needs.